add ind_transform file to include functions
